Write large doubles in scientific notation A candidate fix for issue #40. If a double value will not fit in the allotted field width with fixed `%*.*f` formatting, an attempt is made to reformat it using `%*.*e` scientific notation. Formatted double values are now written with `DBFWriteAttributeDirectly()`, as Shapelib's `DBFWritedDoubleAttribute()` does not support scientific notation and will simply truncate larger-than-width+precision values. Note that Shapelib can *read* numeric values expressed in scientific notation without trouble via the `atof()` function it uses for parsing.
*Important*: the stored value may differ from the provided value due to loss of significant digits when expressed in scientific notation. Effort is made to make maximal use of the available field width to preserve precision. For best accuracy, applications should create sufficiently wide fields to accomodate any possible values.

This issue was noticed when testing the following program to create a copy of the Natural Earth countries layer:

	demos/redim.tcl tests/sample/110m_cultural/ne_110m_admin_0_countries tests/sample/xym/polygonm 0

The billion+ population for some countries such as China and India is encoded in scientific notation in the Natural Earth attribute tables. It can be read ok, but the 10 integer digits are too many to fit in the 9 + . + 6 afforded by the 16.6 field width/precision. With this fix, the program runs successfully and those exceptionally large values are written back out in scientific notation too.